// flog/src/components/commons/buttons/outlined/04/OutlinedButton04.types.ts
export interface IOutlinedButton04Props {
  content: string;
  size: string;
  onClick: () => void;
  disabled?: boolean;
}

export interface IOutlinedButton04UIProps {
  content: string;
  size: string;
  onClick: () => void;
  disabled?: boolean;
}

export interface IOutlinedButton04UIEmotionProps {
  size: string;
  onClick: () => void;
  disabled?: boolean;
}